Struct coco::deque::Deque

struct Deque<T> {
    bottom: AtomicIsize,
    top: AtomicIsize,
    buffer: Atomic<Buffer<T>>,
}

Fields
bottom: AtomicIsize top: AtomicIsize buffer: Atomic<Buffer<T>>
Methods

impl<T> Deque<T>

A work-stealing deque.

fn new() -> Self

Returns a new, empty deque.

fn len(&self) -> usize

Returns the number of elements in the deque.

If used concurrently with other operations, the returned number is just an estimate.

unsafe fn resize(&self, new_cap: usize)

Resizes the buffer with new capacity of new_cap.

fn push(&self, value: T)

Pushes an element onto the bottom of the deque.

fn pop(&self) -> Option<T>

Pops an element from the bottom of the deque.

fn steal(&self) -> Option<T>

Steals an element from the top of the deque.

fn steal_as_worker(&self) -> Option<T>

Steals an element from the top of the deque, but only the worker may call this method.