Function coco::epoch::flush

pub fn flush(pin: &Pin)

Flushes the buffered thread-local garbage.

It is wise to flush the garbage just after passing a very large object to defer_free, so that it isn't sitting in the buffer for a long time.