Struct coco::epoch::garbage::Bag

pub struct Bag {
    len: AtomicUsize,
    objects: [UnsafeCell<(unsafe fn(_: *mut u8, _: usize), *mut u8, usize)>; 64],
    epoch: usize,
    next: Atomic<Bag>,
}

Holds removed objects that will be eventually destroyed.

Fields
len: AtomicUsize

Number of objects in the bag.

objects: [UnsafeCell<(unsafe fn(_: *mut u8, _: usize), *mut u8, usize)>; 64]

Removed objects.

epoch: usize

The global epoch at the moment when this bag got pushed into the queue.

next: Atomic<Bag>

The next bag in the queue.

Methods

impl Bag

pub fn new() -> Self

Returns a new, empty bag.

pub fn is_empty(&self) -> bool

Returns true if the bag is empty.

pub fn try_insert<T>(
&self,
destroy: unsafe fn(_: *mut T, _: usize),
object: *mut T,
count: usize
) -> bool

Attempts to insert a garbage object into the bag and returns true if succeeded.

unsafe fn destroy_all_objects(&self)

Destroys all objects in the bag.

Note: can be called only once!