Struct coco::epoch::garbage::Garbage

pub struct Garbage {
    head: Atomic<Bag>,
    tail: Atomic<Bag>,
    pending: Atomic<Bag>,
}

A garbage queue.

This is where a concurrent data structure can store removed objects for deferred destruction.

Stored garbage objects are first kept in the garbage buffer. When the buffer becomes full, it's objects are flushed into the garbage queue. Flushing can be manually triggered by calling flush.

Some garbage in the queue can be manually collected by calling collect.

Fields
head: Atomic<Bag>

Head of the queue.

tail: Atomic<Bag>

Tail of the queue.

pending: Atomic<Bag>

The next bag that will be pushed into the queue as soon as it gets full.

Methods

impl Garbage

pub fn new() -> Self

Returns a new, empty garbage queue.

fn replace_pending<'p>(
&self,
old: Ptr<'p, Bag>,
pin: &'p Pin
) -> Result<Ptr<'p, Bag>, Ptr<'p, Bag>>

Attempts to compare-and-swap the pending bag old with a new, empty one.

The return value is a result indicating whether the compare-and-swap successfully installed a new bag. On success the new bag is returned. On failure the current more up-to-date pending bag is returned.

pub unsafe fn defer_free<T>(&self, object: *mut T, count: usize, pin: &Pin)

Adds an object that will later be freed.

The specified object is an array allocated at address object and consists of count elements of type T.

This method inserts the object into the garbage buffer. When the buffers becomes full, it's objects are flushed into the garbage queue.

pub unsafe fn defer_drop<T>(&self, object: *mut T, count: usize, pin: &Pin)

Adds an object that will later be dropped and freed.

The specified object is an array allocated at address object and consists of count elements of type T.

This method inserts the object into the garbage buffer. When the buffers becomes full, it's objects are flushed into the garbage queue.

Note: The object must be Send + 'self.

pub unsafe fn defer_destroy<T>(
&self,
destroy: unsafe fn(_: *mut T, _: usize),
object: *mut T,
count: usize,
pin: &Pin
)

Adds an object that will later be destroyed using destroy.

The specified object is an array allocated at address object and consists of count elements of type T.

This method inserts the object into the garbage buffer. When the buffers becomes full, it's objects are flushed into the garbage queue.

Note: The object must be Send + 'self.

pub fn flush(&self, pin: &Pin)

Flushes the buffered garbage.

It is wise to flush the garbage just after passing a very large object to one of the defer_* methods, so that it isn't sitting in the buffer for a long time.

pub fn collect(&self, pin: &Pin)

Collects some garbage from the queue and destroys it.

Generally speaking, it's not necessary to call this method because garbage production already triggers garbage destruction. However, if there are long periods without garbage production, it might be a good idea to call this method from time to time.

This method collects several buffers worth of garbage objects.

fn push(&self, bag: Box<Bag>, pin: &Pin)

Pushes a bag into the queue.

fn try_pop_if<'p, F>(&self, condition: F, pin: &'p Pin) -> Option<&'p Bag> where
F: Fn(&Bag) -> bool,

Attempts to pop a bag from the front of the queue and returns it if condition is met.

If the bag in the front doesn't meet it or if the queue is empty, None is returned.