Struct coco::epoch::Pin

pub struct Pin { /* fields omitted */ }

A witness that the current thread is pinned.

A reference to Pin is proof that the current thread is pinned. Lots of methods that interact with Atomics can safely be called only while the thread is pinned so they often require a reference to Pin.

This data type is inherently bound to the thread that created it, therefore it does not implement Send nor Sync.