Struct coco::epoch::Ptr

pub struct Ptr<'p, T: 'p> { /* fields omitted */ }

A tagged nullable pointer.

Methods

impl<'p, T: 'p> Ptr<'p, T>

pub fn null(tag: usize) -> Self

Returns a null pointer with a tag.

Panics

Panics if the tag doesn't fit into the unused bits of an aligned pointer.

pub unsafe fn from_raw(raw: *mut T, tag: usize) -> Self

Constructs a tagged pointer from a raw pointer and tag.

Panics

Panics if the tag doesn't fit into the unused bits of the pointer, or if the pointer is unaligned.

pub fn is_null(&self) -> bool

Returns true if the pointer is null.

pub fn as_ref(&self) -> Option<&'p T>

Converts the pointer to a reference.

pub fn as_raw(&self) -> *mut T

Converts the pointer to a raw pointer.

pub fn unwrap(&self) -> &'p T

Returns a reference to the pointing object.

Panics

Panics if the pointer is null.

pub fn tag(&self) -> usize

Returns the tag.

pub fn with_tag(&self, tag: usize) -> Self

Constructs a new tagged pointer with a different tag.

Panics

Panics if the tag doesn't fit into the unused bits of the pointer.