Module coco::epoch::thread

Thread synchronization and pinning.

Registration

In order to track all threads in one place, we need some form of thread registration. Every thread has a thread-local so-called "harness" that registers it the first time it is pinned, and unregisters when it exits.

Registered threads are tracked in a global lock-free singly-linked list of thread entries. The head of this list is accessed by calling the participants function.

Thread entries

Thread entries are implemented as the Thread data type. Every entry contains an integer that tells whether the thread is pinned and if so, what was the global epoch at the time it was pinned. Entries also hold a pin counter that aids in periodic global epoch advancement.

Structs

Harness
Holds thread-local data and unregisters the thread when dropped.

Pin
A witness that the current thread is pinned.

Thread
An entry in the linked list of participating threads.

Constants

HARNESS
The thread registration harness.

Functions

defer_free⚠
Stashes away an object that will later be freed.

flush
Flushes the buffered thread-local garbage.

is_pinned
Returns true if the current thread is pinned.

participants
Returns a reference to the head pointer of the list of participating threads.

pin
Pins the current thread.

try_advance
Attempts to advance the global epoch.