Struct coco::epoch::thread::Thread

struct Thread {
    state: AtomicUsize,
    next: Atomic<Thread>,
}

An entry in the linked list of participating threads.

Fields
state: AtomicUsize

The least significant bit is set if the thread is currently pinned. The rest of the bits encode the current epoch.

next: Atomic<Thread>

The next thread in the linked list of participants. If the tag is 1, this entry is deleted and can be unlinked from the list.

Methods

impl Thread

fn set_pinned(&self, _pin: &Pin)

Marks the thread as pinned.

Must not be called if the thread is already pinned!

fn set_unpinned(&self)

Marks the thread as unpinned.

fn register() -> *mut Thread

Registers a thread by adding a new entry to the list of participanting threads.

Returns a pointer to the newly allocated entry.

fn unregister(&self)

Unregisters the thread by marking it's entry as deleted.

This function doesn't physically remove the entry from the linked list, though. That will do a future call to try_advance.